Refactor link checking logic in AffordanceEntity
The link checking logic in the AffordanceEntity class has been refactored to include an additional condition for the "vocab.nerc" link and a new condition for the "zenodo.org" link. This change ensures that the parent element's "mia-extra-properties" attribute is checked for the "nochange" value, preventing the element from being changed.