Refactor link checking logic in AffordanceEntity
Added method where if no type is found for the uri then all triples where the subject is the uri will be gathered. The template is then set to default. This will ensure that whenever you parse a url and get some triples that a _blank node being the type does not take away from the information.

Additionally doi.org has been added to the list of urls that can be parsed